fix(cli): Rename test_analytics.py to analytics.py to avoid wheel validation error
The file `test_analytics.py` was being flagged as a test file during package
validation because it starts with "test_", but it's actually a CLI command for
analyzing test coverage data, not a pytest test file.

Changes:
- Rename src/nhl_scrabble/cli/commands/test_analytics.py -> analytics.py
- Update imports in cli/__init__.py and commands/__init__.py
- Update C901 exemption in pyproject.toml
- Update architecture documentation

The command remains accessible as `nhl-scrabble test-analytics` - only the
module filename changed to prevent false positives in wheel validation.

Fixes GitHub Actions "Package Validation" workflow error.
