+++ src/nhl_scrabble/cli/commands/__init__.py
@@ -6,12 +6,12 @@ in its own module for better organization and maintainability.
 
 from __future__ import annotations
 
+from nhl_scrabble.cli.commands.analytics import test_analytics
 from nhl_scrabble.cli.commands.analyze import analyze
 from nhl_scrabble.cli.commands.dashboard import dashboard
 from nhl_scrabble.cli.commands.interactive_cmd import interactive
 from nhl_scrabble.cli.commands.search import search
 from nhl_scrabble.cli.commands.serve import serve
-from nhl_scrabble.cli.commands.test_analytics import test_analytics
 from nhl_scrabble.cli.commands.watch import watch
 
 __all__ = [
