+++ src/nhl_scrabble/cli/commands/analytics.py
